Replace `% 8` with `& 7`
See https://stackoverflow.com/a/11040718/4760801
for more information. Basically, if the denominator is known
to be a power of two, you can replace `n % m` with `n & (m - 1)`.

The compiler will not actually make this optimization (always). GCC 9.2
seems to be the only one that gets close (yet still not as simple),
but ICC 19.0.1 and CLANG 9.0.0 use division (`idiv`).